Initial work for location and zoom input
