+++ templates/river/widgets/location_and_precision.html
